+++ packages/expo-router/src/__tests__/initialRouteName.test.ios.tsx
@@ -100,21 +100,33 @@ it('push should include (group)/index as an anchor route when using withAnchor',
     '(group)/test': () => null,
   });
 
-  // Initial stale state
+  // Initial complete state
   expect(store.state).toStrictEqual({
+    index: 0,
+    key: expect.any(String),
+    routeNames: ['__root', '+not-found', '_sitemap'],
     routes: [
       {
+        key: expect.any(String),
         name: '__root',
         state: {
+          index: 0,
+          key: expect.any(String),
+          routeNames: ['index', '(group)'],
           routes: [
             {
+              key: expect.any(String),
               name: 'index',
               path: '/',
             },
           ],
+          stale: false,
+          routeKeySeq: expect.any(Number),
         },
       },
     ],
+    stale: false,
+    routeKeySeq: expect.any(Number),
   });
 
   act(() => router.push('/orange', { withAnchor: true }));
@@ -140,7 +152,7 @@ it('push should include (group)/index as an anchor route when using withAnchor',
             {
               key: expect.any(String),
               name: '(group)',
-              params: { initial: false, params: { initial: false }, screen: 'orange' },
+              params: {},
               path: undefined,
               state: {
                 index: 1,
@@ -154,19 +166,23 @@ it('push should include (group)/index as an anchor route when using withAnchor',
                   {
                     key: expect.any(String),
                     name: 'orange',
-                    params: { initial: false },
+                    params: {},
+                    path: '/orange',
                   },
                 ],
                 stale: false,
+                routeKeySeq: expect.any(Number),
               },
             },
           ],
           stale: false,
+          routeKeySeq: expect.any(Number),
           type: 'stack',
         },
       },
     ],
     stale: false,
+    routeKeySeq: expect.any(Number),
     type: 'stack',
   });
 });
@@ -181,21 +197,33 @@ it('push should ignore (group)/index as an initial route if no anchor is specifi
     '(group)/test': () => null,
   });
 
-  // Initial stale state
+  // Initial complete state
   expect(store.state).toStrictEqual({
+    index: 0,
+    key: expect.any(String),
+    routeNames: ['__root', '+not-found', '_sitemap'],
     routes: [
       {
+        key: expect.any(String),
         name: '__root',
         state: {
+          index: 0,
+          key: expect.any(String),
+          routeNames: ['index', '(group)'],
           routes: [
             {
+              key: expect.any(String),
               name: 'index',
               path: '/',
             },
           ],
+          stale: false,
+          routeKeySeq: expect.any(Number),
         },
       },
     ],
+    stale: false,
+    routeKeySeq: expect.any(Number),
   });
 
   act(() => router.push('/orange'));
@@ -221,7 +249,7 @@ it('push should ignore (group)/index as an initial route if no anchor is specifi
             {
               key: expect.any(String),
               name: '(group)',
-              params: { params: {}, screen: 'orange' },
+              params: {},
               path: undefined,
               state: {
                 index: 0,
@@ -232,18 +260,22 @@ it('push should ignore (group)/index as an initial route if no anchor is specifi
                     key: expect.any(String),
                     name: 'orange',
                     params: {},
+                    path: '/orange',
                   },
                 ],
                 stale: false,
+                routeKeySeq: expect.any(Number),
               },
             },
           ],
           stale: false,
+          routeKeySeq: expect.any(Number),
           type: 'stack',
         },
       },
     ],
     stale: false,
+    routeKeySeq: expect.any(Number),
     type: 'stack',
   });
 });