+++ packages/expo-router/src/getLinkingConfig.ts
@@ -6,9 +6,8 @@ import type { State } from './fork/getPathFromState';
 import { getReactNavigationConfig } from './getReactNavigationConfig';
 import { applyRedirects } from './getRoutesRedirects';
 import type { StoreRedirects } from './global-state/router-store';
-import { getInitialURL, getPathFromState, subscribe } from './link/linking';
+import { getInitialURL, getPathFromState, getStateFromPath, subscribe } from './link/linking';
 import type { LinkingOptions } from './react-navigation/native';
-import { getActionFromState } from './react-navigation/native';
 import type { NativeIntent, RequireContext } from './types';
 
 export function getNavigationConfig(
@@ -47,6 +46,7 @@ export function getNavigationConfig(
 
 export type ExpoLinkingOptions<T extends object = Record<string, unknown>> = LinkingOptions<T> & {
   getPathFromState: typeof getPathFromState;
+  getStateFromPath: typeof getStateFromPath;
 };
 
 export type LinkingConfigOptions = {
@@ -110,13 +110,19 @@ export function getLinkingConfig(
           if (typeof initialUrl === 'string') {
             initialUrl = applyRedirects(initialUrl, redirects);
             if (initialUrl && typeof nativeLinking?.redirectSystemPath === 'function') {
-              initialUrl = nativeLinking.redirectSystemPath({ path: initialUrl, initial: true });
+              initialUrl = nativeLinking.redirectSystemPath({
+                path: initialUrl,
+                initial: true,
+              });
             }
           } else if (initialUrl) {
             initialUrl = initialUrl.then((url) => {
               url = applyRedirects(url, redirects);
               if (url && typeof nativeLinking?.redirectSystemPath === 'function') {
-                return nativeLinking.redirectSystemPath({ path: url, initial: true });
+                return nativeLinking.redirectSystemPath({
+                  path: url,
+                  initial: true,
+                });
               }
               return url;
             });
@@ -127,6 +133,7 @@ export function getLinkingConfig(
       return initialUrl;
     },
     subscribe: subscribe(nativeLinking, redirects),
+    getStateFromPath,
     getPathFromState(state: State, options: Parameters<typeof getPathFromState>[1]) {
       return (
         getPathFromState(state, {
@@ -136,9 +143,6 @@ export function getLinkingConfig(
         }) ?? '/'
       );
     },
-    // Add all functions to ensure the types never need to fallback.
-    // This is a convenience for usage in the package.
-    getActionFromState,
   };
 }