+++ packages/expo-router/src/global-state/routeInfoContext.ts
@@ -1,0 +1,7 @@
+'use client';
+
+import { createContext } from 'react';
+
+import { defaultRouteInfo } from './getRouteInfoFromState';
+
+export const RouteInfoContext = createContext(defaultRouteInfo);